backend/requirements.txt
Django==4.2.8
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-filter==23.5
psycopg2-binary==2.9.9
python-dotenv==1.0.0
Pillow==10.1.0

# For PostGIS support
GDAL==3.6.2

# Additional utilities
PyJWT==2.8.0
python-dateutil==2.8.2
